Python/Whizzbuy_Work/DevEnvironment/requirements.txt
aniso8601==1.1.0
awsebcli==3.7.2
botocore==1.3.17
cement==2.4.0
colorama==0.3.3
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
Flask==0.10.1
Flask-Login==0.3.2
flask-mysql==1.3
Flask-MySQLdb==0.2.0
Flask-OpenID==1.2.5
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
gunicorn==19.4.5
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
MarkupSafe==0.23
mysql-python==1.2.5
mysqlclient==1.3.7
pathspec==0.3.3
python-dateutil==2.4.2
python-openid==2.2.5
pytz==2015.7
PyYAML==3.11
requests==2.6.2
six==1.10.0
source==1.2.0
SQLAlchemy==1.0.10
texttable==0.8.4
virtualenv==13.1.2
websocket-client==0.35.0
Werkzeug==0.11.2
wheel==0.26.0
bugsnag
blinker==1.4
newrelic
celery
flask-mail
APScheduler==3.0.0
geolocation-python
js2py
yattag
